feat: теперь Resource использует float вместо int

+++ include/PercentageResource.h
@@ -8,23 +8,23 @@ struct PercentageResource : public Resource {
 private:
     std::string name;
     Texture2D icon;
-    int currentValue;
-    int maxValue;
+    float currentValue;
+    float maxValue;
 
 public:
-    PercentageResource(const std::string& resourceName, const std::string& texturePath, int startValue = 0, int maximumValue = 1000);
+    PercentageResource(const std::string& resourceName, const std::string& texturePath, float startValue = 0, float maximumValue = 1000);
     ~PercentageResource();
 
     std::string GetName() const override { return name; }
     void Draw(const Rectangle& area) const override;
-    int GetValue() const override { return currentValue; }
-    void SetValue(int newValue) override;
+    float GetValue() const override { return currentValue; }
+    void SetValue(float newValue) override;
     std::string GetDisplayText() const override;
 
-    void Add(int value) override;
-    void Subtract(int value) override;
-    void SetMaxValue(int newMaxValue);
-    int GetMaxValue() const { return maxValue; }
+    void Add(float value) override;
+    void Subtract(float value) override;
+    void SetMaxValue(float newMaxValue);
+    float GetMaxValue() const { return maxValue; }
     float GetPercentage() const { return static_cast<float>(currentValue) / maxValue; }
 };
